Parameterized URL as CLI arg
Often when profiling, you want to 'parameterize' on part of the search
URL.  Example, you want to replicate the same request, but with
different ports, or different query params, etc.  This commit adds this
kind of behavior, specified as CLI arguments.  For example, to
parameterize on ports, you could do:
<query input stream> | lein run <sample-size> :port 80 81
Other examples:
<is> | lein run <ss> :query-params '{mode discrete}' '{mode loose}'
Accepted parameter keys:
:protocol, :domain, :port, :path, :query-str, :query-params